pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/latch"]

[project]
name = "latch"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pyyaml",
  "aiohttp",
  "pyotp",
  "qrcode",
  "rich",
  "questionary",
]

[project.scripts]
latch = "latch.cli:main"
latch-server = "latch.server:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
